refactor: extract reused functions into components
